move py35_* tests to original modules
