+++ tests/pool_test.py
@@ -531,3 +531,25 @@ async def test_pool_idle_close(create_pool, start_server, loop):
     await asyncio.sleep(2, loop=loop)
 
     assert (await conn.execute('ping')) == b'PONG'
+
+
+@pytest.mark.run_loop
+async def test_await(create_pool, server, loop):
+    pool = await create_pool(
+        server.tcp_address,
+        minsize=10, loop=loop)
+
+    with await pool as conn:
+        msg = await conn.execute('echo', 'hello')
+        assert msg == b'hello'
+
+
+@pytest.mark.run_loop
+async def test_async_with(create_pool, server, loop):
+    pool = await create_pool(
+        server.tcp_address,
+        minsize=10, loop=loop)
+
+    async with pool.get() as conn:
+        msg = await conn.execute('echo', 'hello')
+        assert msg == b'hello'
